Fix hyperparameters suited to SAMSum dataset. Add CLI arguments to script.
The hyperparameters used by the long-summarization model are too big for SAMSum
dataset and this causes a problem while training and slows down the model a lot.
We conduct a few exploratory experiments on the dataset and fix on these
hyperparameters.

We also added CLI arguments to the run.sh script to make it more versatile while
using in colab.